dev-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

-r requirements.txt

# Unit test:
pytest>=2.4
pytest-cov
mock
requests-mock

# Sphinx:
# Note: The ordereddict package is a backport of collections.OrderedDict
#       to Python 2.6 and earlier. OrderedDict is needed by GitPython, which
#       is needed by sphinx-git.
Sphinx>=1.3
ordereddict ; python_version < '2.7'
GitPython>=2.0.6
sphinx-git

# PyLint:
# Astroid is used by Pylint. Astroid 1.3 and above, and Pylint 1.4
# and above no longer work with Python 2.6, and have been removed
# from Pypi in 2/2016 after being available for some time.
# Therefore, we cannot use Pylint under Python 2.6.
# Also, Pylint does not support Python 3.
astroid ; python_version == '2.7'
pylint ; python_version == '2.7'

# Twine: Needed for uploading to Pypi
twine

# Jupyter Notebook
# Install manually, to where you want to run Jupyter.
# jupyter

# Examples:
PyYAML
